+++ twilight-interactions-derive/src/command/mod.rs
@@ -2,6 +2,7 @@
 
 mod impls;
 
+mod description;
 mod model;
 mod subcommand;