+++ twilight-interactions-derive/src/command/model/create_command.rs
@@ -3,7 +3,7 @@ use quote::{quote, quote_spanned};
 use syn::{spanned::Spanned, DeriveInput, Error, FieldsNamed, Result};
 
 use super::parse::{channel_type, command_option_value, optional, StructField, TypeAttribute};
-use crate::parse::{find_attr, parse_doc};
+use crate::{command::description::get_description, parse::find_attr};
 
 /// Implementation of `CreateCommand` derive macro
 pub fn impl_create_command(input: DeriveInput, fields: Option<FieldsNamed>) -> Result<TokenStream> {
@@ -36,16 +36,18 @@ pub fn impl_create_command(input: DeriveInput, fields: Option<FieldsNamed>) -> R
         ));
     }
 
+    let desc = get_description(
+        &attributes.desc_localizations,
+        &attributes.desc,
+        span,
+        &input.attrs,
+    )?;
+
     let name = match &attributes.name {
         Some(name) => name,
         None => return Err(Error::new(attr_span, "missing required attribute `name`")),
     };
     let name_localizations = localization_field(&attributes.name_localizations);
-    let description = match &attributes.desc {
-        Some(desc) => desc.clone(),
-        None => parse_doc(&input.attrs, span)?,
-    };
-    let description_localizations = localization_field(&attributes.desc_localizations);
     let default_permissions = match &attributes.default_permissions {
         Some(path) => quote! { ::std::option::Option::Some(#path())},
         None => quote! { ::std::option::Option::None },
@@ -73,11 +75,12 @@ pub fn impl_create_command(input: DeriveInput, fields: Option<FieldsNamed>) -> R
 
                 #(#field_options)*
 
+                let desc = #desc;
                 ::twilight_interactions::command::ApplicationCommandData {
                     name: ::std::convert::From::from(#name),
                     name_localizations: #name_localizations,
-                    description: ::std::convert::From::from(#description),
-                    description_localizations: #description_localizations,
+                    description: desc.0,
+                    description_localizations: desc.1,
                     options: command_options,
                     default_member_permissions: #default_permissions,
                     dm_permission: #dm_permission,
@@ -94,13 +97,15 @@ fn field_option(field: &StructField) -> Result<TokenStream> {
     let ty = &field.ty;
     let span = field.span;
 
+    let desc = get_description(
+        &field.attributes.desc_localizations,
+        &field.attributes.desc,
+        field.span,
+        &field.raw_attrs,
+    )?;
+
     let name = field.attributes.name_default(field.ident.to_string());
     let name_localizations = localization_field(&field.attributes.name_localizations);
-    let description = match &field.attributes.desc {
-        Some(desc) => desc.clone(),
-        None => parse_doc(&field.raw_attrs, field.span)?,
-    };
-    let description_localizations = localization_field(&field.attributes.desc_localizations);
     let required = field.kind.required();
     let autocomplete = field.attributes.autocomplete;
     let max_value = command_option_value(field.attributes.max_value);
@@ -116,12 +121,13 @@ fn field_option(field: &StructField) -> Result<TokenStream> {
     };
 
     Ok(quote_spanned! {span=>
+        let desc = #desc;
         command_options.push(<#ty as ::twilight_interactions::command::CreateOption>::create_option(
             ::twilight_interactions::command::internal::CreateOptionData {
                 name: ::std::convert::From::from(#name),
                 name_localizations: #name_localizations,
-                description: ::std::convert::From::from(#description),
-                description_localizations: #description_localizations,
+                description: desc.0,
+                description_localizations: desc.1,
                 required: ::std::option::Option::Some(#required),
                 autocomplete: #autocomplete,
                 data: ::twilight_interactions::command::internal::CommandOptionData {