+++ twilight-interactions/src/command/mod.rs
@@ -39,14 +39,14 @@
 //! and the second tuple element is the localized value.
 //!
 //! ```
-//! use twilight_interactions::command::{CommandModel, CreateCommand, ResolvedUser};
+//! use twilight_interactions::command::{CommandModel, CreateCommand, ResolvedUser, DescriptionLocalizations};
 //!
 //! #[derive(CommandModel, CreateCommand)]
-//! #[command(name = "hello", desc = "Say hello", desc_localizations = "hello_desc")]
+//! #[command(name = "hello", desc_localizations = "hello_desc")]
 //! struct HelloCommand;
 //!
-//! pub fn hello_desc() -> [(&'static str, &'static str); 2] {
-//!     [("fr", "Dis bonjour"), ("de", "Sag Hallo")]
+//! pub fn hello_desc() -> DescriptionLocalizations {
+//!     DescriptionLocalizations::new("Say hello", [("fr", "Dis bonjour"), ("de", "Sag Hallo")])
 //! }
 //! ```
 //!
@@ -84,6 +84,7 @@
 
 mod command_model;
 mod create_command;
+mod localizations;
 
 #[doc(hidden)]
 pub mod internal;
@@ -93,6 +94,7 @@ pub use command_model::{
     ResolvedUser,
 };
 pub use create_command::{ApplicationCommandData, CreateCommand, CreateOption};
+pub use localizations::{DescriptionLocalizations, NameLocalizations};
 #[cfg(feature = "derive")]
 #[cfg_attr(docsrs, doc(cfg(feature = "derive")))]
 pub use twilight_interactions_derive::{CommandModel, CommandOption, CreateCommand, CreateOption};