+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - A basic example bot implementation has been added in the `examples` directory
   of the repository.
+- `DescriptionLocalizations` and `NameLocalizations` structs
+
+### Changed
+- `desc` and `desc_localizations` are now mutually exclusive
+- `desc_localizations` and `name_localizations` must return the
+  `DescriptionLocalizations` and `NameLocalizations` structs,
+  respectively.
 
 ## [0.15.2] - 2023-06-23
 ### Added